+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PCB Print: Workaround for KiCad 9.0.5 change in how holes are plotted when
   a layer is not included (black instead of white).
 - Export_3D: `center` option to `origin` (#871)
+- Compress: Now `dest` also expands %X patterns (can be disabled) (#873)
 
 ### Fixed
 - User.N layer numbering. I.e. pcb_print issues with their numbering (#808)